requirements.txt
asgiref==3.5.2
Django==3.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-pandas==0.6.6
numpy==1.23.3
pandas==1.4.4
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
